Integration test

+++ src/main/java/be/vdab/cinefest/films/FilmController.java
@@ -13,7 +13,7 @@ public class FilmController {
     }
 
     @GetMapping("films/totaalvrijeplaatsen")
-    long findTotaalVrijePlaatsen() {
+    int findTotaalVrijePlaatsen() {
         return filmService.findAantalVrijePlaatsen();
     }
